requirements.txt
flask
python-dotenv
line-bot-sdk
requests
google-generativeai
pytesseract
Pillow
gspread
oauth2client
psycopg2-binary
gunicorn
APScheduler
pytz